The populate() method will add infinitesimal cells, which will then be "grown" slowly to their desired sizes (while continuing to hold their properties and ability to "overlap").

+++ Cells/Healthy.cpp
@@ -9,7 +9,7 @@
 
 #include "Healthy.h"
 
-Healthy::Healthy(double x, double y, float radius)
+Healthy::Healthy(double x, double y, double radius)
         : Cell(x, y, radius)
 {}
